requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
disnake==2.9.1
dnspython==2.6.1
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
isodate==0.6.1
multidict==6.0.5
numpy==1.26.4
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo[encryption]==4.7.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
rsa==4.9
six==1.16.0
toml==0.10.2
twitchAPI==4.2.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
asyncache==0.3.1
flask[async]==3.0.3
gunicorn==22.0.0
waitress==3.0.0